chore: fix `make mocks`
