+++ testutil/mock/types_mock_appmodule.go
@@ -360,40 +360,3 @@ func (mr *MockCoreAppModuleMockRecorder) ValidateGenesis(arg0 interface{}) *gomo
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ValidateGenesis", reflect.TypeOf((*MockCoreAppModule)(nil).ValidateGenesis), arg0)
 }
-
-// MockUpgradeModule is a mock of UpgradeModule interface.
-type MockUpgradeModule struct {
-	*MockCoreAppModule
-	recorder *MockUpgradeModuleMockRecorder
-}
-
-// MockUpgradeModuleMockRecorder is the mock recorder for MockUpgradeModule.
-type MockUpgradeModuleMockRecorder struct {
-	*MockCoreAppModuleMockRecorder
-}
-
-// NewMockUpgradeModule creates a new mock instance.
-func NewMockUpgradeModule(ctrl *gomock.Controller) *MockUpgradeModule {
-	mock := &MockUpgradeModule{
-		MockCoreAppModule: NewMockCoreAppModule(ctrl),
-	}
-	mock.recorder = &MockUpgradeModuleMockRecorder{mock.MockCoreAppModule.recorder}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockUpgradeModule) EXPECT() *MockUpgradeModuleMockRecorder {
-	return m.recorder
-}
-
-// IsUpgradeModule mocks base method.
-func (m *MockUpgradeModule) IsUpgradeModule() {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "IsUpgradeModule")
-}
-
-// IsUpgradeModule indicates an expected call of IsUpgradeModule.
-func (mr *MockUpgradeModuleMockRecorder) IsUpgradeModule() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsUpgradeModule", reflect.TypeOf((*MockUpgradeModule)(nil).IsUpgradeModule))
-}
