+++ testutil/mock/types_module_module.go
@@ -129,6 +129,41 @@ func (mr *MockHasNameMockRecorder) Name() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Name", reflect.TypeOf((*MockHasName)(nil).Name))
 }
 
+// MockUpgradeModule is a mock of UpgradeModule interface.
+type MockUpgradeModule struct {
+	ctrl     *gomock.Controller
+	recorder *MockUpgradeModuleMockRecorder
+}
+
+// MockUpgradeModuleMockRecorder is the mock recorder for MockUpgradeModule.
+type MockUpgradeModuleMockRecorder struct {
+	mock *MockUpgradeModule
+}
+
+// NewMockUpgradeModule creates a new mock instance.
+func NewMockUpgradeModule(ctrl *gomock.Controller) *MockUpgradeModule {
+	mock := &MockUpgradeModule{ctrl: ctrl}
+	mock.recorder = &MockUpgradeModuleMockRecorder{mock}
+	return mock
+}
+
+// EXPECT returns an object that allows the caller to indicate expected use.
+func (m *MockUpgradeModule) EXPECT() *MockUpgradeModuleMockRecorder {
+	return m.recorder
+}
+
+// IsUpgradeModule mocks base method.
+func (m *MockUpgradeModule) IsUpgradeModule() {
+	m.ctrl.T.Helper()
+	m.ctrl.Call(m, "IsUpgradeModule")
+}
+
+// IsUpgradeModule indicates an expected call of IsUpgradeModule.
+func (mr *MockUpgradeModuleMockRecorder) IsUpgradeModule() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsUpgradeModule", reflect.TypeOf((*MockUpgradeModule)(nil).IsUpgradeModule))
+}
+
 // MockHasGenesisBasics is a mock of HasGenesisBasics interface.
 type MockHasGenesisBasics struct {
 	ctrl     *gomock.Controller
